pyproject.toml
[project]
name = "dentobase"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.12.0",
    "alembic>=1.18.2",
    "anthropic>=0.79.0",
    "argon2-cffi>=25.1.0",
    "asyncpg>=0.31.0",
    "chromadb>=1.4.1",
    "einops>=0.8.2",
    "fastapi[standard]>=0.128.0",
    "grandalf>=0.8",
    "greenlet>=3.3.1",
    "huggingface-hub>=0.36.2",
    "ipython>=9.10.0",
    "jose>=1.0.0",
    "langchain>=1.2.7",
    "langchain-anthropic>=1.3.2",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.0",
    "langchain-groq>=1.1.1",
    "langchain-huggingface>=1.2.0",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.7",
    "langchain-text-splitters>=1.1.0",
    "passlib>=1.7.4",
    "pillow>=12.1.0",
    "psycopg2>=2.9.11",
    "pydantic-settings>=2.12.0",
    "pypdf>=6.6.2",
    "python-dotenv>=1.2.1",
    "python-jose>=3.5.0",
    "resend>=2.21.0",
    "sentence-transformers>=5.2.2",
    "sqlalchemy>=2.0.46",
    "timm>=1.0.24",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "transformers>=4.57.6",
    "uvicorn[standard]>=0.40.0",
]
